Split out the explicit compressors.
