+++ document/src/vespa/document/util/compressor.h
@@ -1,12 +1,12 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
-#include <vespa/document/util/compressionconfig.h>
-#include <vespa/vespalib/data/databuffer.h>
+#include "compressionconfig.h"
 #include <vespa/vespalib/util/buffer.h>
 
-namespace document
-{
+namespace vespalib { class DataBuffer; }
+
+namespace document {
 
 class ICompressor
 {
@@ -16,15 +16,6 @@ public:
     virtual bool unprocess(const void * input, size_t inputLen, void * output, size_t & outputLen) = 0;
     virtual size_t adjustProcessLen(uint16_t options, size_t len)   const = 0;
     virtual size_t adjustUnProcessLen(uint16_t options, size_t len) const = 0;
-};
-
-class LZ4Compressor : public ICompressor
-{
-public:
-    bool process(const CompressionConfig& config, const void * input, size_t inputLen, void * output, size_t & outputLen) override;
-    bool unprocess(const void * input, size_t inputLen, void * output, size_t & outputLen) override;
-    size_t adjustProcessLen(uint16_t options, size_t len)   const override;
-    size_t adjustUnProcessLen(uint16_t options, size_t len) const override;
 };
 
 /**
@@ -52,4 +43,3 @@ CompressionConfig::Type compress(const CompressionConfig & compression, const ve
 void decompress(const CompressionConfig::Type & compression, size_t uncompressedLen, const vespalib::ConstBufferRef & org, vespalib::DataBuffer & dest, bool allowSwap);
 
 }
-
