+++ searchlib/src/vespa/searchlib/docstore/chunkformat.cpp
@@ -1,7 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "chunkformats.h"
-#include <vespa/document/util/compressor.h>
+#include <vespa/document/util/lz4compressor.h>
 #include <vespa/vespalib/util/stringfmt.h>
 
 namespace search {
